add config creation to book, simplify example structure
