--- examples/parachain-example/metadata/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "metadata"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-subxt = { path = "../../../subxt" }
