--- examples/parachain-example/metadata/src/lib.rs
@@ -1,2 +1,0 @@
-#[subxt::subxt(runtime_metadata_path = "statemint_metadata.scale")]
-pub mod statemint {}
