+++ examples/parachain-example/statemint_metadata.scale
