+++ subxt/src/book/setup/client.rs
@@ -12,17 +12,15 @@
 //! Both clients are generic over a [`crate::config::Config`] trait, which is the way that we give
 //! the client certain information about how to interact with a node that isn't otherwise available
 //! or possible to include in the node metadata.
+//!
 //! The [`crate::config::Config`] trait mimics the `frame_system::Config` trait and
 //! subxt ships out of the box with two default implementations:
 //!
 //! - [`crate::config::PolkadotConfig`] for talking to Polkadot nodes, and
 //! - [`crate::config::SubstrateConfig`] for talking to generic nodes built with Substrate.
 //!
-//! The latter will generally work in many cases, but will need modifying if the chain you'd like to
-//! connect to has altered any of the details mentioned in [the trait](`crate::config::Config`).
-//!
-//! For a detailed breakdown of how to construct a config for a custom chain, please refer to [the parachain example](../../../../../../examples/parachain-example/target/doc/parachain_example/index.html).
-//! Usually the [`PolkadotConfig`](crate::config::PolkadotConfig) and [`SubstrateConfig`](crate::config::PolkadotConfig) should be sufficient though.
+//! The latter will generally work in many cases, but [may need special customization](super::config) if
+//! the node differs in any of the types the [`Config`](crate::config::Config) trait wants to know about.
 //!
 //! In the case of the [`crate::OnlineClient`], we have a few options to instantiate it:
 //!
